Stable Staff Roster Widget
In addition to Stable Staff Roster Widget, I've updated the registerAccount page to inform the user of password requirements.
