+++ operationportal/lib/Staff/Staff_ClassViewer.dart
@@ -23,10 +23,9 @@ import '../Structs/Choice.dart';
 import '../SuspensionView.dart';
 
 class Staff_ClassViewer_Page extends StatefulWidget {
-  Staff_ClassViewer_Page({Key key, this.user, this.classId}) : super(key: key);
+  Staff_ClassViewer_Page({Key key, this.mClass}) : super(key: key);
 
-  final User user;
-  final String classId;
+  final Class mClass;
 
   @override
   Staff_ClassViewer_State createState() => Staff_ClassViewer_State();
@@ -64,50 +63,25 @@ class Staff_ClassViewer_State extends State<Staff_ClassViewer_Page> {
             "lastTireChange": DateTime,
             "lastMaintenance": DateTime
           */
-          FutureBuilder(
-              future: storage.readToken().then((value) {
-                return RetrieveClassInfo(value, widget.classId, context);
-              }),
-              builder: (BuildContext context, AsyncSnapshot<Class> snapshot) {
-                switch (snapshot.connectionState) {
-                  case ConnectionState.none:
-                    return new Text('Issue Posting Data');
-                  case ConnectionState.waiting:
-                    return new Center(child: new CircularProgressIndicator());
-                  case ConnectionState.active:
-                    return new Text('');
-                  case ConnectionState.done:
-                    if (snapshot.hasError) {
-                      return Center(
-                        child: Text("Unable to Fetch Class Info"),
-                      );
-                    } else {
-                      return Center(
-                          child: Column(
-                            children: <Widget>[
-                              Container(
-                                child: Text("Name:\n${snapshot.data.teacherName}", textAlign: TextAlign.center,
-                                  style: TextStyle(color: Colors.white),),
-                                decoration: new BoxDecoration(
-                                  color: Colors.blue,
-                                  borderRadius: new BorderRadius.all(
-                                      new Radius.circular(20)
-                                  ),
-                                ),
-                                margin: EdgeInsets.only(
-                                  top: 20,
-                                ),
-                                padding: EdgeInsets.all(20),
-                              ),
-                            ],
-                          )
-                      );
-                    }
-                    break;
-                  default:
-                    return null;
-                }
-              }
+          Center(
+              child: Column(
+                children: <Widget>[
+                  Container(
+                    child: Text("Name:\n${widget.mClass.teacherName}", textAlign: TextAlign.center,
+                      style: TextStyle(color: Colors.white),),
+                    decoration: new BoxDecoration(
+                      color: Colors.blue,
+                      borderRadius: new BorderRadius.all(
+                          new Radius.circular(20)
+                      ),
+                    ),
+                    margin: EdgeInsets.only(
+                      top: 20,
+                    ),
+                    padding: EdgeInsets.all(20),
+                  ),
+                ],
+              )
           ),
         ],
       ),
